Enable to use the action only against the changed files in git (#2)
* Fix typo in README.md (#1)

* Exit with status code 1 one the setup fails

The logging was pintpointing the exact root of the problem,
but subsequent execeptions were raised since the code didn't
stop the execution.

* Improve exception handling during setup

It avoids to catch generic Exceptions.

* Add feature to specify list of files via environment instead of searching in directory

This change introduce a new functionality that allows the user
to specify a list of files to process. It is useful when you
have different confluence destination in the same directory
or you want to specify a fixed list of file, eg. the files changed
in the commit.

* Add instructions on how to upload only changed files in README.md

* Fix typo in action.yml

* Improve example of verifying changed file in README.md

* Refine README.md section on how to upload only changed files

The change provides an example to cover all the situation
of pull_requests and pushes, with tip and tricks on how
to make it work flawlessly.